Module cont_arith_code::distribution::gaussian

Structs

Gaussian

Numerically stable Gaussian distribution

TruncatedGaussian

WithReplacement

Static (non-optimal) Gaussian model of a set of integers

WithoutReplacement

Updating (optimal) Gaussian model of a set of integers

Functions

floor_rem

Separate the fractional part of a number of the integer part

lerp

Does a linear interpolation on the area under the curve of the PDF of the standard normal between two points. Does the computation in log-scale so we’re numerically stable in the tails

log_cdf

log-CDF of the standard normal

log_pdf

log-PDF of the standard normal

log_probability

Returns a numerically stable log(abs(cdf(hi) - cdf(lo))) for the standard normal. Should only return log(0.0) if the bounds are the same or if the log-PDF at their mid point is f64::NEG_INFINITY (or almost).

pdf

PDF of the standard normal

quantile

Quantile of the standard normal

quantile_exp

Quantile-exp (inverse of the log-CDF) of the standard normal