Trait cont_arith_code::Model

pub trait Model<T> {
    // Required methods
    fn push(&mut self, s: Index) -> Option<T>;
    fn next_distr(&mut self) -> Box<dyn UnivariateDistribution>;
}

Models produce distributions and consume their resolutions (as i64s) until a value T pops out. Used by both Encoder and Decoder symmetrically.

Required Methods

fn push(&mut self, s: Index) -> Option<T>

Push a value into the model and updates its state. Returns a values once it’s been fully described by a series of indexes.

fn next_distr(&mut self) -> Box<dyn UnivariateDistribution>

Get the distribution for the next symbol from the model.