Trait cont_arith_code::TruncatedDistribution

pub trait TruncatedDistribution {
    // Required methods
    fn quantile(&self, cp: f64) -> (Index, f64);
    fn truncate(&mut self, cp: f64, s: Index, s_rem: f64, bit: bool);
    fn lo(&self) -> Index;
    fn hi(&self) -> Index;

    // Provided method
    fn is_resolved(&self) -> bool { ... }
}

A univariate distribution with a truncated probability mass between a lower- and upper-bound.

Required Methods

fn quantile(&self, cp: f64) -> (Index, f64)

Quantile (inverse CDF) and decompose of the remaining probability mass. Returns whole index (i64) and remainder (f64 $\in$ [0-1]) whose sum has the given cummulative probability in the truncated CDF.

fn truncate(&mut self, cp: f64, s: Index, s_rem: f64, bit: bool)

Split the remaining probability mass with the given bit (false:0:left :: true:1:right) at the given cummulative probability, which we already know splits at index s with remainder s_rem.

fn lo(&self) -> Index

Return the index of the lower-bound of the interval of the truncation.

fn hi(&self) -> Index

Return the index of the upper-bound of the interval of the truncation.

Provided Methods

fn is_resolved(&self) -> bool

True iff self.lo() == self.hi()